requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
duckdb==0.9.2
duckdb_engine==0.11.2
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlglot==22.5.0
sqlglotrs==0.1.2
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
duckdb==0.9.2
duckdb_engine==0.11.2
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlglot==22.5.0
sqlglotrs==0.1.2
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
vercel-kv==0.1.5
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
duckdb==0.9.2
duckdb_engine==0.11.2
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
multidict==6.0.5
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlglot==22.5.0
sqlglotrs==0.1.2
starlette==0.36.3
typing_extensions==4.10.0
upstash-redis==1.0.0
urllib3==2.2.1
uvicorn==0.28.0
yarl==1.9.4
